setup unit tests for Nix
Both for a `nix-build`, where `cabal` executable is not present, but
also just for using Stack with nix integration. With nix integration,
Stack will make a `nix-shell` where `stack` executable is not found.

We use "bare-bones" GHC to solve this. This GHC doesn't exactly respect
our Stack or Cabal configurations, so we need to avoid `cubicbezier`'s
`matrices` dependency so that `Data.Matrix` resolves to `matrix`
package.

Also, we don't error on compiler warnings when compiling test files
through Nix. I don't know if these warnings would show up on a regular
Stack or Cabal config (I'm on NixOS), so it's unaffected for those.

TL;DR: Added testing support for Nix build environments; Cabal and Stack
environments should be unchanged.