Fix last edge of rectangle rendering in reverse order
I found an issue where a rectangle that is supposed to be animated (in
terms of corners) `BL -> BR -> TR -> TL -> cycle back to BL` are instead
rendered as `BL -> BR -> TR -> TL`, and then suddenly cuts to the
completed rectangle, before retracting the line `TL -> BL` from the BL
corner (so, instead of a line dropping from TL down, imagine the full
line retracting from BL up toward TL).

It seems that `LineEnd` doesn't mean "draw a line from the current
position to the initial position" like `Z` does in SVG specifications.
So, I manually added the last edge of a rectangle as a `PathCommand`.

This does change all the tests that depend on animating rectangles. The
only difference is the explicit `v<negative number>` representing the
downstroke. I'm not sure if this is ideal (i.e. if `LineEnd` is intended
to behave like SVG `Z`) but it works.